Move full text match test utils to TestUtils
