+++ common/CMakeLists.txt
@@ -61,7 +61,6 @@ add_library( LiteCoreJNI SHARED
     ${JNI_SRC}/native_c4collection.cc
     ${JNI_SRC}/native_c4database.cc
     ${JNI_SRC}/native_c4document.cc
-    ${JNI_SRC}/native_c4fulltextmatch.cc
     ${JNI_SRC}/native_c4index.cc
     ${JNI_SRC}/native_c4listener.cc
     ${JNI_SRC}/native_c4observer.cc
